feat: implement useStakingPendingTxs hook for managing staking transactions
